Add another version
Instead of the precision score, recall score is considered in this version. However, the results are the same. The graphic of the recall score, compared to a range of thresholds is added to this version. 

In addition, in the previous version, only X_test0 obtain from the train_test_split was used to compute the evaluation score chosen. However, in this version, two samples are tested from another dataset provided separately from the one splitted to train and test the model in the previous version. Thus, the model is tested using two new dataset. The mean of the recall score is 0.89,